feat: refactor gtag, theme effect
