+++ jsconfig.json
@@ -5,6 +5,7 @@
       "@/*": ["./*"],
       "@components/*": ["src/components/*"],
       "@containers/*": ["src/containers/*"],
+      "@hooks/*": ["src/hooks/*"],
       "@libs/*": ["src/libs/*"],
       "@styles/*": ["src/styles/*"]
     }
