--- src/components/Theme.jsx
@@ -1,22 +1,0 @@
-export const getTheme = () => {
-  if (
-    localStorage.theme === 'dark' ||
-    (!('theme' in localStorage) &&
-      window.matchMedia('(prefers-color-scheme: dark)').matches)
-  ) {
-    return 'dark'
-  } else {
-    return 'light'
-  }
-}
-
-function Theme() {
-  if (getTheme() === 'dark') {
-    document.documentElement.classList.add('dark')
-  } else {
-    document.documentElement.classList.remove('dark')
-  }
-  return null
-}
-
-export default Theme
