+++ src/libs/gtag.js
@@ -3,6 +3,7 @@ export const GA_TRACKING_ID = CONFIG.googleAnalytics.config.measurementId
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
 export const pageview = url => {
+  if (typeof window !== 'object') return
   window.gtag('config', GA_TRACKING_ID, {
     page_path: url
   })
@@ -10,6 +11,7 @@ export const pageview = url => {
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/events
 export const event = ({ action, category, label, value }) => {
+  if (typeof window !== 'object') return
   window.gtag('event', action, {
     event_category: category,
     event_label: label,
